Extract LeaveRequestValidator from LeaveRequestService
The leave-request validation (policy rules, day-count computation, and
overlap detection) was tangled into LeaveRequestService alongside the request
lifecycle. Move it into a focused LeaveRequestValidator; the approval workflow
already lives in LeaveApprovalService, so this completes the split.

LeaveRequestService injects the validator: createRequest and submitRequest call
leaveRequestValidator.validate, and the public calculateTotalDays /
getConflictingDates methods (which controllers use) become thin delegates. The
LeaveBalanceService dependency moves to the validator. Behaviour and the REST
surface are unchanged; the moved read methods keep their transaction semantics,
which for the internal self-calls were already the caller's transaction.
